S realizaron ajustes en varios componentes de la UI, viewmodels de inicio
