+++ FinanKey/MauiProgram.cs
@@ -60,7 +60,7 @@ namespace FinanKey
             builder.Services.AddSingleton<ViewModelGasto>();
             builder.Services.AddSingleton<ViewModelIngreso>();
             builder.Services.AddSingleton<ViewModelCuenta>();
-            builder.Services.AddSingleton<ViewModelFinanzas>();
+            builder.Services.AddSingleton<ViewModelInicio>();
             builder.Services.AddSingleton<ViewModelDetalleCuenta>();
             builder.Services.AddSingleton<ViewModelTarjeta>();
             //Registrar servicios con inyección de dependencias contexto de datos
