+++ FinanKey/ViewModels/ViewModelInicio.cs
@@ -1,0 +1,72 @@
+﻿using CommunityToolkit.Mvvm.ComponentModel;
+using CommunityToolkit.Mvvm.Input;
+using FinanKey.Models;
+using FinanKey.Servicios;
+using FinanKey.View;
+using System.Collections.ObjectModel;
+
+namespace FinanKey.ViewModels
+{
+    public partial class ViewModelInicio : ObservableObject
+    {
+        //Inyección de dependencias 
+        public readonly IServicioTarjeta _servicioTarjeta;
+        //Inicializar la lista de cuentas como una colección observable
+        [ObservableProperty]
+        private ObservableCollection<Tarjeta> _listaTarjetas = new();
+        [ObservableProperty]
+        private bool isBusy;
+        [ObservableProperty]
+        private bool _hayTarjetas = false;
+
+        public ViewModelInicio(IServicioTarjeta servicioTarjeta)
+        {
+            //Asignar los servicios a las variables privadas
+            _servicioTarjeta = servicioTarjeta;
+        }
+
+        public async Task CargarTarjetasAsync()
+        {
+            try
+            {
+                var tarjetas = await _servicioTarjeta.ObtenerTarjetasAsync();
+
+                if (tarjetas != null && tarjetas.Count > 0)
+                {
+                    ListaTarjetas = new ObservableCollection<Tarjeta>(tarjetas);
+                    HayTarjetas = !true;
+                }
+                else
+                {
+                    ListaTarjetas = new ObservableCollection<Tarjeta>();
+                    HayTarjetas = !false;
+                }
+            }
+            catch (Exception ex)
+            {
+                await Shell.Current.DisplayAlert("Error", $"Error al cargar tarjetas: {ex.Message}", "OK");
+            }
+        }
+        // // Comando para navegar a la página de detalle de cuenta con la cuenta seleccionada
+        //[RelayCommand]
+        //async Task NavegarADetalleCuenta(Cuenta cuenta)
+        //{
+        //    if (cuenta is null)
+        //        return;
+
+        //    cuentaSeleccionada = cuenta;
+
+        //    await Shell.Current.GoToAsync(
+        //        nameof(DetalleCuentaPage),
+        //        new Dictionary<string, object>
+        //        {
+        //            ["Cuenta"] = cuentaSeleccionada
+        //        });
+        //}
+        [RelayCommand]
+        public async Task NavegarAgregarTarjeta()
+        {
+            await Shell.Current.GoToAsync(nameof(AgregarTarjetaPage));
+        }
+    }
+}
